requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
Flask-SQLAlchemy==2.5.1
requests==2.25.1
python-dotenv==0.17.1
